+++ src/node/iri.js
@@ -94,7 +94,7 @@ class IRI extends Base {
                             reject(err);
                             return;
                         }
-                        this.log('Neighbors removed', peers.map(p => p.getNelsonURI()));
+                        this.log('Neighbors removed:'.red, peers.map(p => p.getNelsonURI()));
                         resolve(peers)
                     });
                 }
@@ -118,7 +118,7 @@ class IRI extends Base {
                     reject(error);
                     return;
                 }
-                this.log('Neighbors added:', data, uris.join(', '));
+                this.log('Neighbors added:'.green, data, uris.join(', '));
                 resolve(peers);
             });
         });
@@ -148,6 +148,24 @@ class IRI extends Base {
                 Array.isArray(neighbors) && neighbors.length
                     ? this.api.removeNeighbors(neighbors.map((n) => `${n.connectionType}://${n.address}`), addNeighbors)
                     : addNeighbors();
+            });
+        });
+    }
+
+    /**
+     * Removes all IRI neighbors.
+     * @returns {Promise}
+     */
+    removeAllNeighbors () {
+        return new Promise((resolve) => {
+            this.api.getNeighbors((error, neighbors) => {
+                if(error) {
+                    return resolve();
+                }
+                if (Array.isArray(neighbors) && neighbors.length) {
+                    return this.api.removeNeighbors(neighbors.map((n) => `${n.connectionType}://${n.address}`), resolve);
+                }
+                resolve();
             });
         });
     }