+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Improved logs
+- Fixed orphaned IRI neighbors
+
 ## [0.1.5] - 2017-12-14
 
 ### Changed